Checks environment variable for artifact download location
As per #177, this change checks the `EMBEDDED_MONGO_ARTIFACTS` environment variable when picking the default location to store downloaded artifacts. This allows the location to be controlled at an environmental level rather than an application level, making it possible to manage usage of a shared continuous integration environment more centrally.